feat: componentize the UploadAvatarFieldWithCrop
