+++ packages/toolkit/src/view/settings/user/UserProfileTab.tsx
@@ -1,12 +1,11 @@
 import * as z from "zod";
 import * as React from "react";
 
-import { Setting, instillUserRoles } from "..";
+import { Setting } from "..";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import {
   Button,
-  Dialog,
   Form,
   Input,
   Switch,
@@ -24,8 +23,7 @@ import {
   useAuthenticatedUser,
 } from "../../../lib";
 import { FormLabel } from "../FormLabel";
-import { LoadingSpin } from "../../../components";
-import AvatarEditor from "react-avatar-editor";
+import { LoadingSpin, UploadAvatarFieldWithCrop } from "../../../components";
 import { useUpdateAuthenticatedUser } from "../../../lib";
 
 export const UserProfileTabSchema = z.object({
@@ -59,13 +57,6 @@ export const UserProfileTab = () => {
   const { amplitudeIsInit } = useAmplitudeCtx();
   const { accessToken, enabledQuery } = useInstillStore(useShallow(selector));
   const { toast } = useToast();
-  const [profileAvatar, setProfileAvatar] = React.useState<
-    string | File | null
-  >(null);
-  const [isOpenProfileAvatar, setIsOpenProfileAvatar] =
-    React.useState<boolean>(false);
-
-  const editorRef = React.useRef<AvatarEditor>(null);
 
   const me = useAuthenticatedUser({
     accessToken,
@@ -115,24 +106,6 @@ export const UserProfileTab = () => {
         toast,
       });
     }
-  }
-
-  function handleSetProfilePicture() {
-    // Save the cropped image as a file or perform any other actions here
-    // For simplicity, let's assume you have a function to handle image upload
-    const croppedImage = editorRef.current
-      ?.getImageScaledToCanvas()
-      ?.toDataURL();
-    setProfileAvatar(croppedImage ?? null);
-    // Close the dialog or perform any other actions
-    setIsOpenProfileAvatar(false);
-  }
-
-  function handleCancelCropProfile() {
-    setIsOpenProfileAvatar(false);
-    setProfileAvatar(null);
-
-    form.resetField("profile.avatar");
   }
 
   return (
@@ -277,65 +250,9 @@ export const UserProfileTab = () => {
               description="This will be displayed on your profile."
             />
             <Setting.TabSectionContent className="gap-y-4">
-              <Form.Field
-                control={form.control}
-                name="profile.avatar"
-                render={({ field }) => {
-                  return (
-                    <Form.Item className="w-full">
-                      <FormLabel
-                        title="Upload your profile test"
-                        optional={true}
-                      />
-                      <Form.Control>
-                        <div>
-                          <label
-                            htmlFor="org-logo-input"
-                            className="flex h-[150px] w-full cursor-pointer flex-col items-center justify-center rounded border border-dashed border-semantic-bg-line bg-semantic-bg-base-bg text-semantic-fg-secondary product-body-text-3-medium"
-                          >
-                            {field.value ? (
-                              <img
-                                src={
-                                  profileAvatar
-                                    ? String(profileAvatar)
-                                    : field.value
-                                }
-                                alt={`${me.data?.name}-profile`}
-                                className="h-[150px] rounded-full object-contain"
-                              />
-                            ) : (
-                              <p>Upload your profile test</p>
-                            )}
-
-                            <Input.Root className="hidden">
-                              <Input.Core
-                                {...field}
-                                id="org-logo-input"
-                                type="file"
-                                accept="images/*"
-                                value={undefined}
-                                onChange={async (e) => {
-                                  const file = e.target.files?.[0];
-                                  if (file) {
-                                    const reader = new FileReader();
-                                    reader.onload = () => {
-                                      const result = reader.result;
-                                      field.onChange(result);
-                                      setProfileAvatar(String(result));
-                                    };
-                                    reader.readAsDataURL(file);
-                                    setIsOpenProfileAvatar(true);
-                                  }
-                                }}
-                              />
-                            </Input.Root>
-                          </label>
-                        </div>
-                      </Form.Control>
-                      <Form.Message />
-                    </Form.Item>
-                  );
-                }}
+              <UploadAvatarFieldWithCrop
+                fieldName="profile.avatar"
+                form={form}
               />
             </Setting.TabSectionContent>
           </Setting.TabSectionRoot>
@@ -444,43 +361,6 @@ export const UserProfileTab = () => {
           </div>
         </form>
       </Form.Root>
-      <Dialog.Root open={isOpenProfileAvatar}>
-        <Dialog.Content className="!w-[400px]">
-          <Dialog.Header>
-            <Dialog.Title>Crop your new profile picture</Dialog.Title>
-          </Dialog.Header>
-          <div className="flex items-center justify-center">
-            {profileAvatar ? (
-              <AvatarEditor
-                ref={editorRef}
-                image={profileAvatar}
-                width={300}
-                height={300}
-                border={20}
-                borderRadius={9999}
-                color={[248, 249, 252]}
-                scale={1}
-              />
-            ) : null}
-          </div>
-          <div className="flex flex-row justify-between gap-x-4">
-            <Button
-              onClick={() => handleSetProfilePicture()}
-              variant="primary"
-              className="w-full"
-            >
-              Set new profile picture
-            </Button>
-            <Button
-              onClick={() => handleCancelCropProfile()}
-              variant="secondaryGrey"
-              className="w-full"
-            >
-              Cancel
-            </Button>
-          </div>
-        </Dialog.Content>
-      </Dialog.Root>
     </Setting.TabRoot>
   );
 };
